chore: refactor/address TODOs
