+++ src/bulkUtils.ts
@@ -5,16 +5,23 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
+import { Writable } from 'node:stream';
+import * as fs from 'node:fs';
+import { EOL } from 'node:os';
+import { Record as SfRecord } from '@jsforce/jsforce-node';
 import type {
   JobInfoV2,
   IngestJobV2Results,
   IngestJobV2SuccessfulResults,
   IngestJobV2FailedResults,
   IngestJobV2UnprocessedRecords,
+  QueryJobV2,
+  QueryJobInfoV2,
 } from '@jsforce/jsforce-node/lib/api/bulk2.js';
+import { Parsable } from '@jsforce/jsforce-node/lib/record-stream.js';
 
 import type { Schema } from '@jsforce/jsforce-node';
-import { Connection, Messages } from '@salesforce/core';
+import { Connection, Logger, Messages } from '@salesforce/core';
 import { IngestJobV2 } from '@jsforce/jsforce-node/lib/api/bulk2.js';
 import { SfCommand, Spinner } from '@salesforce/sf-plugins-core';
 import { Duration } from '@salesforce/kit';
@@ -174,3 +181,82 @@ export const remainingTime =
   (now: number) =>
   (endWaitTime?: number): number =>
     Math.max((endWaitTime ?? now) - now, 0);
+
+export enum ColumnDelimiter {
+  BACKQUOTE = '`',
+  CARET = '^',
+  COMMA = ',',
+  PIPE = '|',
+  SEMICOLON = ';',
+  TAB = '	',
+}
+
+export type ColumnDelimiterKeys = keyof typeof ColumnDelimiter;
+
+export async function getQueryStream(
+  queryJob: QueryJobV2<Schema>,
+  columnDelimiter: ColumnDelimiterKeys,
+  logger: Logger
+): Promise<[Parsable, QueryJobInfoV2]> {
+  const recordStream = new Parsable();
+  const dataStream = recordStream.stream('csv', {
+    delimiter: ColumnDelimiter[columnDelimiter],
+  });
+
+  let jobInfo: QueryJobInfoV2 | undefined;
+
+  try {
+    queryJob.on('jobComplete', (completedJob: QueryJobInfoV2) => {
+      jobInfo = completedJob;
+    });
+    await queryJob.poll();
+
+    const queryRecordsStream = await queryJob.result().then((s) => s.stream());
+    queryRecordsStream.pipe(dataStream);
+  } catch (error) {
+    const err = error as Error;
+    logger.error(`query job failed due to: ${err.message}`);
+
+    if (err.name !== 'JobPollingTimeoutError') {
+      // fires off one last attempt to clean up and ignores the result | error
+      queryJob.delete().catch((ignored: Error) => ignored);
+    }
+
+    throw err;
+  }
+  if (!jobInfo) {
+    throw messages.createError('error.noJobInfo');
+  }
+
+  return [recordStream, jobInfo];
+}
+
+export class JsonWritable extends Writable {
+  private recordsQty: number;
+  private recordsWritten = 0;
+  private filePath: fs.PathLike;
+  private fileStream!: fs.WriteStream;
+
+  public constructor(filePath: fs.PathLike, recordsQty: number) {
+    super({ objectMode: true });
+    this.recordsQty = recordsQty;
+    this.filePath = filePath;
+  }
+
+  public _construct(callback: () => void): void {
+    this.fileStream = fs.createWriteStream(this.filePath);
+    this.fileStream.write(`[${EOL}`);
+    callback();
+  }
+
+  public _write(chunk: SfRecord, encoding: BufferEncoding, callback: () => void): void {
+    if (this.recordsQty - 1 === this.recordsWritten) {
+      // last record, close JSON array
+      this.fileStream.write(`  ${JSON.stringify(chunk)}${EOL}]`, encoding, callback);
+      this.fileStream.end();
+    } else {
+      this.fileStream.write(`  ${JSON.stringify(chunk)},${EOL}`, encoding, callback);
+      this.recordsWritten++;
+    }
+  }
+}
